Re-organise file structure to prepare for redux implementation
