+++ src/Components/Main/Cards/TopicCards/TopicListCard/TopicListCard.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from '@reach/router';
-import ncnLogo from '../../../../../ncnlogo-rb.svg';
+import ncnLogo from '../../../../../assets/global/images/ncnlogo-rb.svg';
 import './TopicListCard.css';
 
 const TopicListCard = ({ topic }) => {
