+++ src/Components/Main/Cards/UserCards/TopUsersCard/TopUsersCard.css
@@ -6,7 +6,7 @@
 }
 
 .top-users-cover-image {
-  background-image: url('../../../../../banner-bg-red.png');
+  background-image: url('../../../../../assets/global/images/banner-bg-red.png');
   background-position: -10px -10px;
   background-size: 200%;
   background-color: #bd1732;
