+++ src/Components/Main/Cards/UserCards/UserProfileCard/UserProfileCard.css
@@ -19,7 +19,7 @@
 }
 
 .user-profile-cover-image {
-  background-image: url('../../../../../banner-bg-red.png');
+  background-image: url('../../../../../assets/global/images/banner-bg-red.png');
   background-position: -10px -10px;
   background-size: 200%;
   background-color: #bd1732;
