+++ src/index.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { WindowProvider } from './Components/Context/WindowProvider';
-import { UserSettingsProvider } from './Components/Context/UserSettingsProvider';
-import { SidebarProvider } from './Components/Context/SidebarProvider';
+import { WindowProvider } from './components/Context/WindowProvider';
+import { UserSettingsProvider } from './components/Context/UserSettingsProvider';
+import { SidebarProvider } from './components/Context/SidebarProvider';
 import './index.css';
-import App from './Components/App/App';
+import App from './components/App/App';
 
 ReactDOM.render(
   <UserSettingsProvider>
